\name{Proportional}
\alias{Proportional}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Proportional Rule
}
\description{
   This function calculates how to distribute a given endowment by the Proportional rule.
}
\usage{
Proportional(E, C, Names = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{E}{
     Endowment
}
  \item{C}{
     Claims of the agents
}
  \item{Names}{
     Labels of the agents
}
}


\author{
Sebastian Cano-Berlanga <cano.berlanga@gmail.com>
}

\note{
In order to calculate the rule properly, input the claims of the agents in ascending order.
}




