\name{summary.ClaimsRule}
\alias{summary.ClaimsRule}


\title{
Summary Method for ClaimsRule Objects
}

\usage{
\method{summary}{ClaimsRule}(object, ...)
}

\arguments{
  \item{object}{an object of class \code{"ClaimsRule"}}
  \item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}

\description{
  \code{\link{summary}} method for class \code{"ClaimsRule"}.
}