\name{summary.ClaimsRules}
\alias{summary.ClaimsRules}
\title{
Summary methods for a ClaimsRules Object
}
\description{
Summary methods for a ClaimsRules Object
}

\usage{
\method{summary}{ClaimsRules}(object, ...)
}

\arguments{
\item{object}{A \code{ClaimsRules} object}
\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}