\name{summary.Game}
\alias{summary.Game}

\title{
Summary methods for a Game Object
}
\description{
Summary methods for a Game Object
}

\usage{
\method{summary}{Game}(object, ...)
}

\arguments{
\item{object}{A \code{Game} object}
\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}