\name{summary.Nucleolus}
\alias{summary.Nucleolus}

\title{
Summary methods for a Nucleolus Object
}
\description{
Summary methods for a Nucleolus Object
}
\usage{
\method{summary}{Nucleolus}(object, ...)
}

\arguments{
\item{object}{A \code{Nucleolus} object}
\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}