\name{summary.ShapleyShubik}
\alias{summary.ShapleyShubik}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary methods for a ShapleyShubik Object
}
\description{
Summary methods for a ShapleyShubik Object
}

\usage{
\method{summary}{ShapleyShubik}(object, ...)
}

\arguments{
\item{object}{A \code{ShapleyShubik} object}
\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}