\name{summary.ShapleyValue}
\alias{summary.ShapleyValue}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary methods for a ShapleyValue Object
}
\description{
Prints the summary of the  Shapley values solution for a given game.
}

\usage{
\method{summary}{ShapleyValue}(object, ...)
}

\arguments{
\item{object}{A \code{ShapleyValue} object}
\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}